Fix/리뷰 이미지 파일 중복, 별점 평균 오류 수정

+++ src/main/java/com/example/tripKo/domain/place/dto/response/info/FestivalResponse.java
@@ -17,7 +17,7 @@ public class FestivalResponse {
 
   private Long id;
   private String name;
-  private Float averageScore;
+  private double averageScore;
   private String mainImage;
   private List<Content> contents;
   private String address;
